1.0 版本，大体完成，需要测试~

+++ v2-test/lib/adaptor.js
@@ -44,7 +44,7 @@ function bindInintConf(mw) {
   mw.config = mw.conf = getDefaultConf();
   mw.initConfig = mw.createServerByConf = function(conf) {
     if (hadInit) {
-      throw new Error('禁止调用多次 inintConfig 方法');
+      throw new Error('禁止调用多次 inintConfig || createServerByConf 方法');
     }
     hadInit = true;
 
@@ -55,42 +55,50 @@ function bindInintConf(mw) {
 }
 
 function bindUtils(mw) {
-  // Builder, SimpleBuilder, request, require1
   mw.Builder = require('./builder/builder');
   mw.SimpleBuilder = require('./builder/simpleBuilder');
-  mw.patBuilder = '';
   mw.nunjucksBuilder = mw.jinjaBuilder = require('./builder/jinja/builder');
   mw.patBuilder = require('./builder/pat/builder');
 
   mw.require1 = require('./common/require1');
+  mw.combine = require('./common/combine');
   mw.request = require('./common/request');
+  mw.proxy = require('http-proxy-middleware');
+  mw.util = util;
 
+  // 在静态目录，寻找静态资源，如果找不到，则开始找外部域名目录
   mw.requestStatic = function(req, res, next) {
     // 查找静态文件，有问题，需要改为从该目录，寻找文件，如果文件存在，则返回，否则，忽略之~
     // 有些奇怪的地址，类似: a.js?xxx ; //b.js，这些都是无法正确识别的
     let url = (typeof req === 'string' ? req : req.url).replace(/[#?].*$/, '').replace(/^\/{2,}/, '/').replace(/^\/+/, '');
     let filepath = path.join(mw.conf.STATIC_TEMPORARY_DIR, './' + url);
+
     if (fs.existsSync(filepath)) {
-      sendFile(res, filepath);
+      sendFile(res, filepath, mw.conf.CODE);
     } else {
-      // 否则，去寻找是否有外部绝对路径，如果有，则去绝对路径下载资源，下载之后，保存到临时目录，然后再走一遍 mw.requestStatic(req, res, next);
+      // 寻找是否有外部绝对路径，如果有，则去绝对路径下载资源，下载之后，保存到临时目录，然后返回客户端
       const dirUrls = mw.conf.STATIC_SOURCE_DIRS.filter(url => util.isHttpURI(url));
       downloadAndSaveImage(dirUrls, url, mw.conf.STATIC_TEMPORARY_DIR, (err, filepath) => {
         if (err) {
           next();
         } else {
-          sendFile(res, filepath);
+          sendFile(res, filepath, mw.conf.CODE);
         }
       });
     }
   };
 }
 
+// 下载及保存文件
+// @param {Array} [dirUrls] 外部目录链接列表
+// @param {String} [filename] 文件名字
+// @param {String} [savedir] 需要保存的目录
+// @param {Function} [callback] 结果回调，callback(error, filepath);
 function downloadAndSaveImage(dirUrls, filename, savedir, callback) {
   const dir = dirUrls.shift();
   if (dir) {
     const url = (/\/$/.test(dir) ? dir : dir + '/') + (/^\//.test(filename) ? filename.replace(/^\/*/, '') : filename);
-    console.log(chalk.green('正在请求外部资源:' + url));
+    console.log(chalk.cyan('download:' + url));
     download(url).then(
       (data) => {
         const filepath = path.join(savedir, './' + filename);
@@ -107,11 +115,25 @@ function downloadAndSaveImage(dirUrls, filename, savedir, callback) {
   }
 }
 
-function sendFile(res, filepath) {
-  // TODO 如果是脚本、或者样式，发现编码不正确，应该修正之~~~~
-  res.sendFile(filepath);
+// 根据编码，发送文件
+function sendFile(res, filepath, code) {
+  // 如果是脚本、或者样式，发现编码不正确，应该修正之~~~~
+  const extname = path.extname(filepath).slice(1);
+  if (fs.existsSync(filepath) && (extname == 'js' || extname == 'css')) {
+    // 进行 gbk 和 utf8 转码
+    const codes = ['utf8', 'gbk'];
+    code && codes.push(code);
+    const content = util.readFile(filepath, codes);
+    res.type(extname).send(content);
+  } else {
+    res.sendFile(filepath);
+  }
 }
 
+// 配置对象，转为为数组
+//  { from: '', to: '' }      => { from: '', to: '' }
+//  './xxx'                   => { from: './xxx', to: '' }
+//  { './x': '', './y': '' }  => [ { from: './x', to: '' }, { from: './y', to: '' } ]
 const confObject2Array = function(obj) {
   const type = util.type(obj);
   if (type === 'object') {
@@ -159,6 +181,11 @@ class OldMockWeb {
 
     // 监听数据文件
     this.watchMockData();
+
+    // 清空临时目录
+    if (this.conf.clean) {
+      this.cleanAfterExit();
+    }
   }
 
   initByConf() {
@@ -211,10 +238,6 @@ class OldMockWeb {
   copyResourceAndWatch() {
     const context = this;
     const conf = this.conf;
-    const list = [].concat(conf.TEMPLATE_SOURCE_DIRS, conf.STATIC_SOURCE_DIRS)
-      .filter(filepath =>
-        !util.isHttpURI(filepath)
-      );
 
     // 已经监听过的，不要重复监听了
     if (!this._resouceWatchMap) {
@@ -243,21 +266,30 @@ class OldMockWeb {
             const filepathRelative = path.relative(dirParent, filepath);
             fs.copySync(filepath, path.join(to, filepathRelative));
           } else {
-            fs.copySync(from, to);
+            const dirParent = from;
+            const filepathRelative = path.relative(dirParent, filepath);
+            fs.copySync(filepath, path.join(to, filepathRelative));
           }
           context.reload();
         }
 
+        if (!is_glob_expr && fs.existsSync(from)) {
+          fs.copySync(from, to);
+        }
+
         // 监听当前文件
-        chokidar.watch(from, { awaitWriteFinish: { stabilityThreshold: 100, pollInterval: 20 } })
-          .on('add', (filepath) => {
+        chokidar.watch(from, { ignoreInitial: is_glob_expr ? false : true })
+          .on('add', function(filepath) {
+            if (is_glob_expr) {
+              console.log(chalk.green(`copy from: ${filepath}`));
+            }
             copyAndReload(filepath);
           })
-          .on('change', (filepath) => {
+          .on('change', function(filepath) {
             console.log('file change');
             copyAndReload(filepath);
           })
-          .on('unlink', (filepath) => {
+          .on('unlink', function(filepath) {
             if (is_glob_expr) {
               const dirParent = globParent(from);
               const filepathRelative = path.relative(dirParent, filepath);
@@ -324,6 +356,25 @@ class OldMockWeb {
     }, 50);
   }
 
+  cleanAfterExit() {
+    const context = this;
+    process.on('SIGINT', function() {
+      console.log('在正清空临时目录...');
+      // 清空临时目录
+      try {
+        if (context.conf.clean) {
+          fs.removeSync(context.conf.TEMPORARY_DIR);
+        }
+      } catch (e) {
+        // nothing~
+      } finally {
+        process.exit(0);
+      }
+    });
+
+    context.cleanAfterExit = function() {};
+  }
+
   setStatic(url, dirname) {
     if (this.server) {
       this.server.setStatic(url, dirname);
@@ -334,7 +385,13 @@ class OldMockWeb {
 
   start() {
     if (this.server) {
-      this.server.start();
+      this.server.start(() => {
+        if (this.conf.openBrowser) {
+          const ip = util.getIps()[0];
+          const page = typeof this.conf.openBrowser === 'string' ? this.conf.openBrowser : '';
+          util.openBrowser(`http://${ip}:${this.conf.port}/${page}`);
+        }
+      });
     }
   }
 }
